requirements.txt
Flask==1.0.2
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
numpy==1.16.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
redis==3.2.1
requests==2.21.0
SQLAlchemy==1.3.3
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xlrd==1.2.0
